Add React with create-react-app
